整理

+++ src/test/java/cn/zyj/tunnel/pathfind/PathCost.java
@@ -2,13 +2,13 @@ package cn.zyj.tunnel.pathfind;
 
 public class PathCost {
 
-    public double toStart;
-    public double toTarget;
-    public double totalCost;
+    public int toStart;
+    public int toTarget;
+    public int totalCost;
 
     public Vec point;
 
-    public PathCost(double toStart, double toTarget, Vec point) {
+    public PathCost(int toStart, int toTarget, Vec point) {
         this.toStart = toStart;
         this.toTarget = toTarget;
         this.totalCost = toStart + toTarget;
