+++ src/test/java/cn/zyj/tunnel/pathfind/Vec.java
@@ -59,8 +59,8 @@ public class Vec {
         return new Vec(data);
     }
 
-    public double distance() {
-        double distance = 0;
+    public int distance() {
+        int distance = 0;
         for (int i = 0; i < data.length; i++) {
             distance += Math.abs(data[i]);
         }
